Fix GUI passing into game
Don't need to call build since it's now called in the constructor. Made it a private function instead. Will need an update or update_score method to update the scores of each team. Build would always keep adding to team num_players for each team, so it can only be called once (in the constructor).